refactor presentation code
