+++ conscious/conscious/MentalStateViewController.swift
@@ -21,7 +21,7 @@ class MentalStateViewController: UIViewController, UIViewControllerTransitioning
     @IBOutlet weak var mentalStateCursorView: UIView!
     
     var mentalStateView: MentalStateView?
-    var animator: MyAnimator?
+    var animator: PresentationAnimator?
     var newPos: CGPoint?
     var startPos: CGPoint?
     var newX: CGFloat?
@@ -30,7 +30,7 @@ class MentalStateViewController: UIViewController, UIViewControllerTransitioning
     var dy: Float!
     var gridCenter: CGPoint?
     var sectors: [Sector]? = []
-    var numberOfSections: Int!
+    var numberOfSections: Int = 4
     var deltaAngle: Float?
     var originalColor: UIColor = UIColor(hexString: "#FAC54B")
     let joyColor: UIColor = UIColor.yellowColor()
@@ -47,7 +47,7 @@ class MentalStateViewController: UIViewController, UIViewControllerTransitioning
         super.viewDidLoad()
         // setup Grid
         setupGrid()
-        createSectors(4)
+        createSectors()
         if (second != nil)  {
             questionLabel.text = "How do you feel now?"
         }
@@ -74,7 +74,7 @@ class MentalStateViewController: UIViewController, UIViewControllerTransitioning
         mentalStateCursorView.clipsToBounds = true
     }
     
-    func createSectors(numberOfSections: Int){
+    func createSectors(){
         let angleSize: CGFloat = CGFloat(2 * M_PI / Double(numberOfSections))
         var mid: CGFloat = 0
         for var i = 0; i < numberOfSections; i++ {
@@ -225,7 +225,7 @@ class MentalStateViewController: UIViewController, UIViewControllerTransitioning
     }
     
     func presentationControllerForPresentedViewController(presented: UIViewController, presentingViewController presenting: UIViewController, sourceViewController source: UIViewController) -> UIPresentationController? {
-        return MyPresentation(presentedViewController: presented, presentingViewController: presenting)
+        return MentalStatePresentation(presentedViewController: presented, presentingViewController: presenting)
     }
     
     /*
